+++ src/Models/ProductVariant.php
@@ -11,9 +11,9 @@ use Hirtz\Shopify\Modules\ModuleTrait;
 use Hirtz\Skeleton\Behaviors\TrailBehavior;
 use Hirtz\Skeleton\Db\ActiveRecord;
 use Hirtz\Skeleton\Db\I18nActiveQuery;
-use Hirtz\Skeleton\Models\Interfaces\AdminRouteInterface;
 use Hirtz\Skeleton\Models\Interfaces\TrailModelInterface;
 use Hirtz\Skeleton\Models\Interfaces\TranslationInterface;
+use Hirtz\Skeleton\Models\Traits\AdminModelTrait;
 use Hirtz\Skeleton\Models\Traits\I18nAttributesTrait;
 use Hirtz\Skeleton\Models\Traits\TrailModelTrait;
 use Hirtz\Skeleton\Models\Traits\TranslationTrait;
@@ -49,8 +49,9 @@ use yii\db\ActiveQuery;
  *
  * @property ProductImage|null $image {@see static::getImage()}
  */
-class ProductVariant extends ActiveRecord implements AdminRouteInterface, TrailModelInterface, TranslationInterface
+class ProductVariant extends ActiveRecord implements TrailModelInterface, TranslationInterface
 {
+    use AdminModelTrait;
     use I18nAttributesTrait;
     use TranslationTrait;
     use ModuleTrait;
@@ -146,19 +147,7 @@ class ProductVariant extends ActiveRecord implements AdminRouteInterface, TrailM
         ]);
     }
 
-    public function getTrailModelName(): string
-    {
-        if ($this->id) {
-            return $this->getI18nAttribute('name') ?: Yii::t('skeleton', 'COMMON_MODEL_ID', [
-                'model' => $this->getTrailModelType(),
-                'id' => $this->id,
-            ]);
-        }
-
-        return $this->getTrailModelType();
-    }
-
-    public function getTrailModelType(): string
+    public function getAdminType(): string
     {
         return Yii::t('shopify', 'COMMON_VARIANT');
     }